fix: Correct node analysis bugs
Fixes two major issues with the node analysis.
The first was a floating point issue, where floating point rounding errors prevented valid nodes from joining with lines.
The second is an issue with T/Y nodes, where the base of the T/Y is joined to the middle of a line, and not on a point, and is therefore missed.